Allow BEL to terminate OSC sequences during setup.
This is a partial cherry-pick of commit de66b0313a23d04 from main,
which allows all the forms of OSC sequence termination to apply in the
preliminary states as well as OSC_STRING.

The reporting user only mentioned the case of OSC 112 BEL, and not the
various forms of ST. So the former is actually known to be occurring
in the wild, and is also the least complicated part of the full patch
on main. Therefore I think this part is worthwhile and reasonably safe
to cherry-pick to 0.77 just before a release, whereas I'd be
uncomfortable making the rest of the changes at this late stage.