GTK: remove 'entrysig' in struct uctrl.
The only *use* of it was removed in commit 6a743399b03e3d6, where
instead of blocking the GTK signal that caused a string to be
overwritten, I switched to making a temporary copy of the string. But
I didn't notice that the declaration and assignments could be cleaned
up too.